Renamed auth::spi to auth::authenticator
