+++ src/auth/authenticator.rs
